Print spatialrd output

Preliminary function, styling with e.g. kable and kableExtra has to be done by the user individually. You could also just use the package of your choice to print out columns of the output from spatialrd.

printspatialrd(SpatialRDoutput)

Arguments

SpatialRDoutput

output file from the spatialrd function

Value

A table with results from the spatialrd function

Examples

points_samp.sf <- sf::st_sample(polygon_full, 1000) # create points
# make it an sf object bc st_sample only created the geometry list-column (sfc):
points_samp.sf <- sf::st_sf(points_samp.sf)
# add a unique ID to each observation:
points_samp.sf$id <- 1:nrow(points_samp.sf)
# assign treatment:
points_samp.sf$treated <- assign_treated(points_samp.sf, polygon_treated, id = "id")
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries
# first we define a variable for the number of "treated" and control
NTr <- length(points_samp.sf$id[points_samp.sf$treated == 1])
NCo <- length(points_samp.sf$id[points_samp.sf$treated == 0])
# the treated areas get a 10 percentage point higher literacy rate
points_samp.sf$education[points_samp.sf$treated == 1] <- 0.7
points_samp.sf$education[points_samp.sf$treated == 0] <- 0.6
# and we add some noise, otherwise we would obtain regression coeffictions with no standard errors
points_samp.sf$education[points_samp.sf$treated == 1] <- rnorm(NTr, mean = 0, sd = .1) +
  points_samp.sf$education[points_samp.sf$treated == 1]
points_samp.sf$education[points_samp.sf$treated == 0] <- rnorm(NCo, mean = 0, sd = .1) +
  points_samp.sf$education[points_samp.sf$treated == 0]

# create distance to cutoff
points_samp.sf$dist2cutoff <- as.numeric(sf::st_distance(points_samp.sf, cut_off))

points_samp.sf$distrunning <- points_samp.sf$dist2cutoff
# give the non-treated one's a negative score
points_samp.sf$distrunning[points_samp.sf$treated == 0] <- -1 *
 points_samp.sf$distrunning[points_samp.sf$treated == 0]

# create borderpoints
borderpoints.sf <- discretise_border(cutoff = cut_off, n = 10)
borderpoints.sf$id <- 1:nrow(borderpoints.sf)

# finally, carry out estimation alongside the boundary:
results <- spatialrd(y = "education", data = points_samp.sf, cutoff.points = borderpoints.sf,
treated = "treated", minobs = 20, spatial.object = FALSE)
printspatialrd(results)
#>    Point    Estimate    SE_Conv     SE_Rob    p_Conv     p_Rob Ntr       Nco
#> 1      1  0.11000000 0.05000000 0.06000000 0.0200000 0.0700000  85  74.00000
#> 2      2  0.13000000 0.05000000 0.06000000 0.0100000 0.0400000 135  68.00000
#> 3      3  0.13000000 0.05000000 0.06000000 0.0100000 0.0200000  71  37.00000
#> 4      4  0.12000000 0.07000000 0.09000000 0.0800000 0.1600000  72 104.00000
#> 5      6  0.00000000 0.09000000 0.10000000 0.9800000 0.8700000  36  62.00000
#> 6      7  0.16000000 0.06000000 0.08000000 0.0100000 0.0300000  89  75.00000
#> 7      8 -0.02000000 0.08000000 0.10000000 0.7900000 0.7000000  65  31.00000
#> 8      9  0.01000000 0.07000000 0.09000000 0.8700000 0.8800000 100  40.00000
#> 9     10  0.09000000 0.07000000 0.09000000 0.2300000 0.4100000 184  52.00000
#> 10  Mean  0.08111111 0.06555556 0.08111111 0.3333333 0.3533333  93  60.33333
#>        bw_l     bw_r CI_Conv_l CI_Conv_u    CI_Rob_l  CI_Rob_u
#> 1  14.90000 14.90000      0.02 0.2000000 -0.01000000 0.2200000
#> 2  18.80000 18.80000      0.03 0.2200000  0.01000000 0.2400000
#> 3  10.40000 10.40000      0.03 0.2300000  0.02000000 0.2700000
#> 4  13.50000 13.50000     -0.02 0.2600000 -0.05000000 0.2900000
#> 5  11.00000 11.00000     -0.17 0.1700000 -0.22000000 0.1900000
#> 6  14.60000 14.60000      0.03 0.2800000  0.02000000 0.3200000
#> 7  10.80000 10.80000     -0.19 0.1400000 -0.24000000 0.1600000
#> 8  13.90000 13.90000     -0.13 0.1500000 -0.18000000 0.1600000
#> 9  21.20000 21.20000     -0.05 0.2300000 -0.10000000 0.2500000
#> 10 14.34444 14.34444     -0.05 0.2088889 -0.08333333 0.2333333