Print spatialrd output

Preliminary function, styling with e.g. kable and kableExtra has to be done by the user individually. You could also just use the package of your choice to print out columns of the output from spatialrd.

printspatialrd(SpatialRDoutput)

Arguments

SpatialRDoutput

output file from the spatialrd function

Value

A table with results from the spatialrd function

Examples

points_samp.sf <- sf::st_sample(polygon_full, 1000) # create points
# make it an sf object bc st_sample only created the geometry list-column (sfc):
points_samp.sf <- sf::st_sf(points_samp.sf)
# add a unique ID to each observation:
points_samp.sf$id <- 1:nrow(points_samp.sf)
# assign treatment:
points_samp.sf$treated <- assign_treated(points_samp.sf, polygon_treated, id = "id")
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries
# first we define a variable for the number of "treated" and control
NTr <- length(points_samp.sf$id[points_samp.sf$treated == 1])
NCo <- length(points_samp.sf$id[points_samp.sf$treated == 0])
# the treated areas get a 10 percentage point higher literacy rate
points_samp.sf$education[points_samp.sf$treated == 1] <- 0.7
points_samp.sf$education[points_samp.sf$treated == 0] <- 0.6
# and we add some noise, otherwise we would obtain regression coeffictions with no standard errors
points_samp.sf$education[points_samp.sf$treated == 1] <- rnorm(NTr, mean = 0, sd = .1) +
  points_samp.sf$education[points_samp.sf$treated == 1]
points_samp.sf$education[points_samp.sf$treated == 0] <- rnorm(NCo, mean = 0, sd = .1) +
  points_samp.sf$education[points_samp.sf$treated == 0]

# create distance to cutoff
points_samp.sf$dist2cutoff <- as.numeric(sf::st_distance(points_samp.sf, cut_off))

points_samp.sf$distrunning <- points_samp.sf$dist2cutoff
# give the non-treated one's a negative score
points_samp.sf$distrunning[points_samp.sf$treated == 0] <- -1 *
 points_samp.sf$distrunning[points_samp.sf$treated == 0]

# create borderpoints
borderpoints.sf <- discretise_border(cutoff = cut_off, n = 10)
borderpoints.sf$id <- 1:nrow(borderpoints.sf)

# finally, carry out estimation alongside the boundary:
results <- spatialrd(y = "education", data = points_samp.sf, cutoff.points = borderpoints.sf,
treated = "treated", minobs = 20, spatial.object = FALSE)
printspatialrd(results)
#>    Point Estimate SE_Conv SE_Rob p_Conv p_Rob   Ntr  Nco  bw_l  bw_r CI_Conv_l
#> 1      1    0.100   0.040  0.040   0.00 0.030 107.0 83.0 18.40 18.40     0.030
#> 2      2    0.070   0.110  0.130   0.53 0.620  76.0 38.0 13.60 13.60    -0.150
#> 3      3    0.080   0.050  0.050   0.07 0.160  56.0 46.0 11.30 11.30    -0.010
#> 4      4    0.060   0.050  0.060   0.23 0.460  37.0 47.0 10.40 10.40    -0.040
#> 5      5    0.060   0.050  0.060   0.26 0.370  76.0 38.0 12.30 12.30    -0.040
#> 6      6   -0.060   0.070  0.080   0.45 0.320  27.0 49.0 10.90 10.90    -0.200
#> 7      7    0.140   0.040  0.050   0.00 0.000  55.0 62.0 11.90 11.90     0.060
#> 8      8    0.080   0.040  0.050   0.04 0.120 135.0 72.0 15.90 15.90     0.000
#> 9      9   -0.060   0.090  0.110   0.52 0.440  80.0 43.0 12.10 12.10    -0.230
#> 10    10    0.070   0.040  0.050   0.10 0.240 257.0 84.0 24.30 24.30    -0.010
#> 11  Mean    0.054   0.058  0.068   0.22 0.276  90.6 56.2 14.11 14.11    -0.059
#>    CI_Conv_u CI_Rob_l CI_Rob_u
#> 1      0.170    0.010    0.190
#> 2      0.300   -0.190    0.320
#> 3      0.170   -0.030    0.180
#> 4      0.150   -0.070    0.160
#> 5      0.150   -0.060    0.170
#> 6      0.090   -0.250    0.080
#> 7      0.230    0.050    0.250
#> 8      0.160   -0.020    0.170
#> 9      0.110   -0.290    0.130
#> 10     0.150   -0.040    0.160
#> 11     0.168   -0.089    0.181