Print spatialrd output

Preliminary function, styling with e.g. kable and kableExtra has to be done by the user individually. You could also just use the package of your choice to print out columns of the output from spatialrd.

printspatialrd(SpatialRDoutput)

Arguments

SpatialRDoutput

output file from the spatialrd function

Value

A table with results from the spatialrd function

Examples

points_samp.sf <- sf::st_sample(polygon_full, 1000) # create points
# make it an sf object bc st_sample only created the geometry list-column (sfc):
points_samp.sf <- sf::st_sf(points_samp.sf)
# add a unique ID to each observation:
points_samp.sf$id <- 1:nrow(points_samp.sf)
# assign treatment:
points_samp.sf$treated <- assign_treated(points_samp.sf, polygon_treated, id = "id")
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries
# first we define a variable for the number of "treated" and control
NTr <- length(points_samp.sf$id[points_samp.sf$treated == 1])
NCo <- length(points_samp.sf$id[points_samp.sf$treated == 0])
# the treated areas get a 10 percentage point higher literacy rate
points_samp.sf$education[points_samp.sf$treated == 1] <- 0.7
points_samp.sf$education[points_samp.sf$treated == 0] <- 0.6
# and we add some noise, otherwise we would obtain regression coeffictions with no standard errors
points_samp.sf$education[points_samp.sf$treated == 1] <- rnorm(NTr, mean = 0, sd = .1) +
  points_samp.sf$education[points_samp.sf$treated == 1]
points_samp.sf$education[points_samp.sf$treated == 0] <- rnorm(NCo, mean = 0, sd = .1) +
  points_samp.sf$education[points_samp.sf$treated == 0]

# create distance to cutoff
points_samp.sf$dist2cutoff <- as.numeric(sf::st_distance(points_samp.sf, cut_off))

points_samp.sf$distrunning <- points_samp.sf$dist2cutoff
# give the non-treated one's a negative score
points_samp.sf$distrunning[points_samp.sf$treated == 0] <- -1 *
 points_samp.sf$distrunning[points_samp.sf$treated == 0]

# create borderpoints
borderpoints.sf <- discretise_border(cutoff = cut_off, n = 10)
borderpoints.sf$id <- 1:nrow(borderpoints.sf)

# finally, carry out estimation alongside the boundary:
results <- spatialrd(y = "education", data = points_samp.sf, cutoff.points = borderpoints.sf,
treated = "treated", minobs = 20, spatial.object = FALSE)
printspatialrd(results)
#>    Point Estimate SE_Conv SE_Rob p_Conv p_Rob   Ntr Nco  bw_l  bw_r CI_Conv_l
#> 1      1    0.150    0.06  0.070  0.010 0.010  73.0  68 13.70 13.70     0.040
#> 2      2    0.110    0.06  0.070  0.060 0.090  79.0  33 13.80 13.80    -0.010
#> 3      3    0.080    0.06  0.070  0.180 0.230  56.0  46 11.90 11.90    -0.040
#> 4      4    0.250    0.09  0.100  0.000 0.010  21.0  35  8.60  8.60     0.080
#> 5      5    0.140    0.05  0.060  0.010 0.020  93.0  56 13.30 13.30     0.040
#> 6      6    0.150    0.10  0.120  0.130 0.230  42.0  77 12.30 12.30    -0.040
#> 7      7    0.150    0.08  0.100  0.070 0.110  32.0  43  9.20  9.20    -0.010
#> 8      8    0.050    0.07  0.090  0.460 0.680  58.0  33 10.40 10.40    -0.080
#> 9      9    0.220    0.07  0.080  0.000 0.000  76.0  45 11.40 11.40     0.080
#> 10    10    0.090    0.06  0.070  0.110 0.250 117.0  54 15.00 15.00    -0.020
#> 11  Mean    0.139    0.07  0.083  0.103 0.163  64.7  49 11.96 11.96     0.004
#>    CI_Conv_u CI_Rob_l CI_Rob_u
#> 1      0.270    0.040    0.310
#> 2      0.230   -0.020    0.250
#> 3      0.200   -0.050    0.230
#> 4      0.420    0.080    0.460
#> 5      0.230    0.020    0.260
#> 6      0.340   -0.100    0.390
#> 7      0.300   -0.030    0.340
#> 8      0.190   -0.130    0.200
#> 9      0.350    0.080    0.400
#> 10     0.200   -0.060    0.210
#> 11     0.273   -0.017    0.305