non-parametric Spatial RD / GRD

This function loops over all boundary points and locally estimates a non-parametric RD (local linear regression as usual) using the rdrobust function from the rdrobust package from Calonico, Cattaneo, Titiunik (2014). It takes in the discretized cutoff point file (the RDcutoff linestring chopped into parts by the discretise_border function) and the sf object (which essentially is just a conventional data.frame with a geometry() column) containing all the observations (treated and untreated). The treated dummy variable has to be assigned before (potentially with assign_treated) and be part of the sf object as a column.

Usage

spatialrd(
  y,
  data,
  cutoff.points,
  treated,
  minobs = 50,
  bwfix_m = NA,
  sample = FALSE,
  samplesize = NA,
  sparse.exclusion = FALSE,
  print.msg = FALSE,
  store.CIs = FALSE,
  spatial.object = TRUE,
  RATestvec = NA,
  ...
)

Arguments

y

The name of the dependent variable in the points frame in the form of a string

data

sf data.frame with points that describe the observations

cutoff.points

sf object of borderpoints (provided by user or obtained with discretise_border)

treated

column that contains the treated dummy (as string)

minobs

the minimum amount of observations in each estimation for the point estimate to be included (default is 50)

bwfix_m

fixed bandwidth in meters (in case you want to impose one yourself)

sample

draw a random sample of points T/F

samplesize

if random, how many points

sparse.exclusion

in case we want to try to exclude sparse border points before the estimation (should reduce warnings)

print.msg

set to TRUE if you want to receive printed info

store.CIs

set TRUE of confidence intervals should be stored

spatial.object

return a spatial object (needed if you want to plot the point estimates on a map)?

RATestvec

vector with strings of basline covariates in the data object in case the RATest by Canay, Kamat (2018) is desired

...

in addition you can use all options in rdrobust

Value

a data.frame or spatial data.frame (sf object) in case spatial.object = T (default)

Details

This function nests rdrobust. All its options (aside from running variable x and cutoff c) are available here as well (e.g. bw selection, cluster level, kernel, weights). Check the documentation in the rdrobust package for details. (bandwidth selection default in rdrobust is bwselect = 'mserd')

To visualise the output table, use printspatialrd or plotspatialrd.

References

Calonico, Cattaneo and Titiunik (2014): Robust Nonparametric Confidence Intervals for Regression-Discontinuity Designs, Econometrica 82(6): 2295-2326.

Examples

if (FALSE) results <- spatialrd(y = "education", data = points_samp.sf, cutoff.points = borderpoints.sf,
treated = "treated", minobs = 10, spatial.object = F)